Revert "Merge dev"
This reverts commit f5f95d4d084dfff8f6e93e6ffc9f7fef71b354e8, reversing
changes made to 1ea569bbd6b70bce94c5558a0cda46c057348add.
