+++ source/georef/georef.h
@@ -157,6 +157,17 @@ public:
 
     void sort_house_numbers();
 
+#ifdef _DEBUG_DIJKSTRA_QUANTUM_
+    template <typename Stream, typename G>
+    void print(Stream& stream, const G& g) {
+        for (auto& edge : this->edges) {
+            stream << std::setprecision(16) << "LINESTRING(" << g[edge.first].coord.lon() << " "
+                   << g[edge.first].coord.lat() << ", " << g[edge.second].coord.lon() << " "
+                   << g[edge.second].coord.lat() << ")" << std::endl;
+        }
+    }
+#endif
+
 private:
     nt::GeographicalCoord get_geographical_coord(const int, const Graph&) const;
     nt::GeographicalCoord extrapol_geographical_coord(int, const Graph&) const;
